refactor: rewrite cjsToEsm.cjs to equivalent in ECMAScript notation (and rename to esm.mjs)
